+++ grammars/level10-Additions.lark
@@ -1,5 +1,5 @@
-command: print | ifs elses? | ask | for_loop | repeat_list | assign_list | sleep | list_access_var | assign | comment
+command: print | repeat | repeat_list | ifs elses? | turtle | ask | assign_list | sleep | list_access_var | assign | comment
 
-//new : for loop
-for_loop: _FOR _SPACE (NAME | var) _SPACE _IN _SPACE _RANGE _SPACE (INT | var) _SPACE _TO _SPACE (INT | var) _EOL (_SPACE command) (_EOL _SPACE command)* _EOL _END_BLOCK
+repeat_list: _FOR _SPACE var _SPACE _IN _SPACE var _EOL (_SPACE command) (_EOL _SPACE command)* _EOL _END_BLOCK
+
