+++ grammars/level11-Additions.lark
@@ -1,17 +1,5 @@
+command: print | ifs elses? | ask | for_loop | repeat_list | assign_list | sleep | list_access_var | assign | comment
 
-//assigning a string now must be done with quotes
-assign: var _SPACE _IS _SPACE sum | var _SPACE _IS _SPACE text_in_quotes
-assign_list: var _SPACE _IS _SPACE (text_in_quotes|NUMBER) ((_COMMA _SPACE|_COMMA) (text_in_quotes|NUMBER))+
+//new : for loop
+for_loop: _FOR _SPACE (NAME | var) _SPACE _IN _SPACE _RANGE _SPACE (INT | var) _SPACE _TO _SPACE (INT | var) _EOL (_SPACE command) (_EOL _SPACE command)* _EOL _END_BLOCK
 
-?atom: NUMBER | var //unsupported numbers are gone, we can now allow floats with NUMBER
-
-//conditions with strings now also use quotes:
-condition: (equality_check|in_list_check) (_SPACE _AND _SPACE condition)*
-
-equality_check: (var | text_in_quotes | NUMBER) _SPACE _IS _SPACE (var | text_in_quotes | NUMBER)
-in_list_check: (var | text_in_quotes | NUMBER) _SPACE _IN _SPACE var
-
-
-invalid: "Supercalifragilisticexpialidocious" //invalid node should be deleted but this probably never matches anything :D
-textwithspaces: /(?:[^\n,' ]| (?!else))+/ -> text //anything can be parsed except for a newline and a comma for list separators, and now single quotes
-text_in_quotes: _SINGLE_QUOTE textwithspaces _SINGLE_QUOTE
