+++ grammars/level4-Additions.lark
@@ -1,28 +1,24 @@
-program: _EOL* (command | invalid) (_SPACE)* (_EOL+ (command | invalid) (_SPACE)*)* _EOL* //lines may end on spaces and might be separated by many newlines
-command: print | ifelse | ifs | ask | turtle | assign_list | sleep | list_access_var | assign | print_no_quotes | comment
-//placing assign after print means print is will print 'is' and print is Felienne will print 'is Felienne'
+command: print | ask | turtle | assign_list | sleep | ask_no_quotes| assign | print_no_quotes | comment | invalid //catch all at the end
 
-list_access_var: var _SPACE _IS _SPACE var _SPACE _AT _SPACE (INT | random)
-assign_list: var _SPACE _IS _SPACE textwithspaces (_COMMA _SPACE? textwithspaces)+
-assign: var _SPACE _IS _SPACE textwithspaces
-invalid: textwithoutspaces (_SPACE textwithspaces)?
+// we need a separate rule for a var when used in a print argument
+// it parses the same, but should not be counted towards lookup table creation
+var_access : NAME
 
-print_no_quotes: _PRINT _SPACE (textwithoutspaces | list_access | var) (_SPACE (textwithoutspaces | list_access | var))*  -> print_nq
+print: _PRINT (_SPACE _print_argument)? -> print
+print_no_quotes: _PRINT _SPACE text -> print_nq
+ask: var _SPACE _IS _SPACE _ASK (_SPACE _print_argument)?
+ask_no_quotes: var _SPACE _IS _SPACE _ASK _SPACE (_SPACE| list_access | textwithoutspaces | punctuation)*  -> print_nq
 
-// new commands for level 4
-ifelse: _IF _SPACE condition (_SPACE|_EOL*) command (_SPACE|_EOL*) _ELSE (_SPACE|_EOL*) command
-ifs: _IF _SPACE condition (_SPACE|_EOL*) command //'if' cannot be used in Python, hence the name of the rule is 'ifs'
-condition: (equality_check|in_list_check) (_SPACE _AND _SPACE condition)*
+_print_argument: (_SPACE | list_access | quoted_text | var_access)*
 
-equality_check: textwithoutspaces _SPACE _IS _SPACE textwithoutspaces //TODO FH nov 2021: not super pretty that this is textwithoutquotes for both a var and also a textual constant, level 11 handles this nicer now, could be done here too
-in_list_check: textwithoutspaces _SPACE _IN _SPACE var
+//anything can be parsed except for spaces, and a newlines and commas for list separators
+textwithoutspaces: /([^\r\n, *+-\/]+)/ -> text
 
-textwithspaces: /(?:[^\n, ]| (?!else))+/ -> text //anything can be parsed except for a newline and a comma for list separators
-//a space is allowed, of course but it may not be followed by an else. The part " (?!else))" means space not followed by (negative look ahead) else
-//That is because allowing else in strings leads to issue #303
-textwithoutspaces: /(?:[^\n, ei]|e(?!lse)|i(?!f))+/ -> text //anything can be parsed except for spaces (plus: a newline and a comma for list separators)
-//the part e(?!lse)|i(?!f)) means e not followed by lse, and i not followed by f
-// this is because allowing else and if in invalid leads to ambiguity in the grammar                                            
+//anything can be parsed except for a newline and a comma for list separators
+//punctuation does not need to be treated differently anymore
+text: /([^\n,]+)/ -> text
 
+quoted_text_no_escape: /'([^']*)'/ -> text //simply all between quotes should this be used at earlier levels?
+quoted_text: /'((?:[^\\']|\\.)*)'/ -> text //text can be between single quotes, but quotes may be escaped with \
