+++ grammars/level6-Additions.lark
@@ -1,4 +1,8 @@
-command: print | repeat | ifelse | ifs | turtle | ask | assign_list | sleep | list_access_var | assign | comment
+print: _PRINT (_SPACE (quoted_text | list_access | var_access | sum) (_SPACE (quoted_text | list_access | var_access | sum))*)?
+assign: var _SPACE _IS _SPACE sum | var _SPACE _IS _SPACE textwithoutspaces
+?sum: product | sum _SPACE* _ADD _SPACE* product -> addition | sum _SPACE* _SUBTRACT _SPACE* product -> substraction
+?product: atom | product _SPACE* _MULTIPLY _SPACE* atom -> multiplication | product _SPACE* _DIVIDE _SPACE* atom -> division
+?atom: INT | var_access | unsupported_number //TODO: means we cannot assign strings with spaces? would we want that?
+unsupported_number: /([-+]?(\d+[\.,]\d+))/ -> unsupported_number
 
-//new in level 6
-repeat: _REPEAT _SPACE (INT | var) _SPACE _TIMES _SPACE command
+textwithoutspaces: /(?:[^\n, *+\-\/ei]|e(?!lse)|i(?!f))+/ -> text //new cause in level 5 calculation elements need to be escaped too
