+++ grammars/level9-Additions.lark
@@ -1,5 +1,2 @@
-command: print | repeat | repeat_list | ifs elses? | turtle | ask | assign_list | sleep | list_access_var | assign | comment
-
-repeat_list: _FOR _SPACE var _SPACE _IN _SPACE var _EOL (_SPACE command) (_EOL _SPACE command)* _EOL _END_BLOCK
-
-
+// no additions to the grammar because we now allow nested blocks
+// which is done in the transpiler
