+++ tests/test_level_04.py
@@ -3,111 +3,227 @@ import textwrap
 from test_level_01 import HedyTester
 
 class TestsLevel4(HedyTester):
-  level=4
+  level = 4
 
-  # test/command order: ['print', 'ask', 'is', 'if', 'turn', 'forward']
+  # tests should be ordered as follows:
+  # * commands in the order of hedy.py for level 3: ['print', 'ask', 'is', 'turn', 'forward'],
+  # * combined tests
+  # * markup tests
+  # * negative tests (inc. negative & multilevel)
 
-  # print & ask -> no changes, covered by tests of earlier levels
+  # test name conventions are like this:
+  # * single keyword positive tests are just keyword or keyword_special_case
+  # * multi keyword positive tests are keyword1_keywords_2
+  # * negative tests should be
+  # * situation_gives_exception
 
-  # is
-  def test_assign_list_access(self):
+
+  # print tests
+  def test_print(self):
     code = textwrap.dedent("""\
-    dieren is Hond, Kat, Kangoeroe
-    dier is dieren at random
-    print dier""")
+    print 'hallo wereld!'""")
 
     result = hedy.transpile(code, self.level)
 
     expected = textwrap.dedent("""\
-    dieren = ['Hond', 'Kat', 'Kangoeroe']
-    dier=random.choice(dieren)
-    print(f'{dier}')""")
+    print(f'hallo wereld!')""")
 
     self.assertEqual(expected, result.code)
     self.assertEqual(False, result.has_turtle)
-    self.assertIn(HedyTester.run_code(result), ['Hond', 'Kat', 'Kangoeroe'])
-  def test_assign_list_multiple_spaces(self):
+  def test_print_comma(self):
     code = textwrap.dedent("""\
-    dieren is Hond,  Kat,       Kangoeroe
-    dier is dieren at random
-    print dier""")
-
-    result = hedy.transpile(code, self.level)
+    naam is Hedy
+    print 'ik heet ,'""")
+    expected = textwrap.dedent("""\
+    naam = 'Hedy'
+    print(f'ik heet ,')""")
+    self.multi_level_tester(
+      code=code,
+      max_level=10,
+      expected=expected,
+      test_name=self.name(),
+      extra_check_function=self.is_not_turtle()
+    )
+  def test_print_two_spaces(self):
+    code = "print        'hallo!'"
 
     expected = textwrap.dedent("""\
-    dieren = ['Hond', 'Kat', 'Kangoeroe']
-    dier=random.choice(dieren)
-    print(f'{dier}')""")
-
-    self.assertEqual(expected, result.code)
-    self.assertEqual(False, result.has_turtle)
-  def test_assign_single_quote(self):
-    code = """message is 'Hello welcome to Hedy.'"""
-    expected = "message = '\\'Hello welcome to Hedy.\\''"
-
-    result = hedy.transpile(code, self.level)
-    self.assertEqual(expected, result.code)
-    self.assertEqual(False, result.has_turtle)
-
-  # if
-  def test_allow_space_after_else_line(self):
-    #this code has a space at the end of line 2
-    code = textwrap.dedent("""\
-    a is 2
-    if a is 1 print a
-    else print 'nee'""")
-
-    result = hedy.transpile(code, self.level)
-
-    expected = textwrap.dedent("""\
-    a = '2'
-    if a == '1':
-      print(f'{a}')
-    else:
-      print(f'nee')""")
-
-    self.assertEqual(expected, result.code)
-  def test_ifelse_should_go_before_assign(self):
-    code = textwrap.dedent("""\
-    kleur is geel
-    if kleur is groen antwoord is ok else antwoord is stom
-    print antwoord""")
-    expected = textwrap.dedent("""\
-      kleur = 'geel'
-      if kleur == 'groen':
-        antwoord = 'ok'
-      else:
-        antwoord = 'stom'
-      print(f'{antwoord}')""")
+    print(f'hallo!')""")
 
     self.multi_level_tester(
+      code=code,
       max_level=4,
+      expected=expected,
+      test_name=self.name(),
+      extra_check_function=self.is_not_turtle()
+    )
+  def test_print_with_slashes(self):
+    code = "print 'Welcome to \\'"
+    result = hedy.transpile(code, self.level)
+
+    expected = textwrap.dedent("""\
+    print(f'Welcome to \\\\')""")
+
+    self.assertEqual(expected, result.code)
+    self.assertEqual(False, result.has_turtle)
+
+    expected_output = HedyTester.run_code(result)
+    self.assertEqual("Welcome to \\", expected_output)
+
+  # ask
+  def test_assign_print(self):
+    code = textwrap.dedent("""\
+    naam is Hedy
+    print 'ik heet' naam""")
+
+    expected = textwrap.dedent("""\
+    naam = 'Hedy'
+    print(f'ik heet{naam}')""")
+
+    self.multi_level_tester(
+      max_level=10,
       code=code,
       expected=expected,
       extra_check_function=self.is_not_turtle(),
       test_name=self.name()
     )
 
-  def test_identifies_backtick_inside_conditional(self):
-    self.assertRaises(hedy.exceptions.UnquotedTextException, lambda: hedy.transpile("if 1 is 1 print `yay!` else print `nay`", self.level))
-
-
-  # turn forward
-  # no new tests, covered by lower levels.
-
-  # combined tests
-  def test_turn_forward(self):
-    result = hedy.transpile("forward 50\nturn\nforward 100", self.level)
-    expected = textwrap.dedent("""\
-    t.forward(50)
-    time.sleep(0.1)
-    t.right(90)
-    t.forward(100)
-    time.sleep(0.1)""")
-    self.assertEqual(expected, result.code)
-    self.assertEqual(True, result.has_turtle)
-  def test_ask_print(self):
+  # issue #745
+  def test_print_list(self):
     code = textwrap.dedent("""\
+        plaatsen is een stad, een  dorp, een strand
+        print 'test' plaatsen""")
+
+    self.multi_level_tester(
+      code=code,
+      max_level=10,
+      exception=hedy.exceptions.InvalidArgumentTypeException,
+      test_name=self.name()
+    )
+
+  def test_ask_Spanish(self):
+    code = textwrap.dedent("""\
+    color is ask 'Cuál es tu color favorito?'""")
+    expected = textwrap.dedent("""\
+    color = input('Cuál es tu color favorito?')""")
+    self.multi_level_tester(
+      max_level=10,
+      code=code,
+      expected=expected,
+      extra_check_function=self.is_not_turtle(),
+      test_name=self.name()
+    )
+  def test_ask_without_quotes(self):
+    code = textwrap.dedent("""
+    ding is kleur
+    kleur is ask Wat is je lievelingskleur'
+    print 'Jouw favoriet is dus ' kleur""")
+
+    with self.assertRaises(hedy.exceptions.UnquotedTextException) as context:
+      result = hedy.transpile(code, self.level)
+
+    self.assertEqual('Unquoted Text', context.exception.error_code)  # hier moet nog we een andere foutmelding komen!
+
+  # is - assign tests
+  def test_assign_underscore(self):
+    code = textwrap.dedent("""\
+    voor_naam is Hedy
+    print 'ik heet '""")
+
+    result = hedy.transpile(code, self.level)
+
+    expected = textwrap.dedent("""\
+    voor_naam = 'Hedy'
+    print(f'ik heet ')""")
+
+    self.assertEqual(expected, result.code)
+    self.assertEqual(False, result.has_turtle)
+  def test_assign_bengali(self):
+    hashed_var = hedy.hash_var("নাম")
+
+    code = textwrap.dedent("""\
+    নাম is হেডি""")
+
+    result = hedy.transpile(code, self.level)
+
+    expected = textwrap.dedent(f"""\
+    {hashed_var} = 'হেডি'""")
+
+    self.assertEqual(expected, result.code)
+    self.assertEqual(False, result.has_turtle)
+  def test_assign_Python_keyword(self):
+    hashed_var = hedy.hash_var("for")
+
+    code = textwrap.dedent("""\
+    for is Hedy
+    print 'ik heet ' for """)
+
+    result = hedy.transpile(code, self.level)
+
+    expected = textwrap.dedent("""\
+    vd55669822f1a8cf72ec1911e462a54eb = 'Hedy'
+    print(f'ik heet {vd55669822f1a8cf72ec1911e462a54eb}')""")
+
+    self.assertEqual(expected, result.code)
+    self.assertEqual(False, result.has_turtle)
+
+  # negative tests
+  def test_print_without_quotes(self):
+    with self.assertRaises(hedy.exceptions.UnquotedTextException) as context:
+      result = hedy.transpile("print felienne 123", self.level)
+
+    self.assertEqual('Unquoted Text', context.exception.error_code)  # hier moet nog we een andere foutmelding komen!
+
+  #combined tests
+  def test_assign_print_bengali(self):
+    hashed_var = hedy.hash_var("নাম")
+    self.assertEqual('veb9b5c786e8cde0910df4197f630ee75', hashed_var)
+
+    code = textwrap.dedent("""\
+    নাম is হেডি
+    print 'আমার নাম is ' নাম """)
+
+    result = hedy.transpile(code, self.level)
+
+    expected = textwrap.dedent("""\
+    veb9b5c786e8cde0910df4197f630ee75 = 'হেডি'
+    print(f'আমার নাম is {veb9b5c786e8cde0910df4197f630ee75}')""")
+
+    self.assertEqual(expected, result.code)
+  def test_assign_print_chinese(self):
+    hashed_var = hedy.hash_var("你好世界")
+    self.assertEqual('v65396ee4aad0b4f17aacd1c6112ee364', hashed_var)
+
+    code = textwrap.dedent("""\
+    你好世界 is 你好世界
+    print 你好世界""")
+
+    result = hedy.transpile(code, self.level)
+
+    expected = textwrap.dedent("""\
+    v65396ee4aad0b4f17aacd1c6112ee364 = '你好世界'
+    print(f'{v65396ee4aad0b4f17aacd1c6112ee364}')""")
+
+    self.assertEqual(expected, result.code)
+
+  def test_print_list_var_random(self):
+
+    code = textwrap.dedent("""\
+    dieren is Hond, Kat, Kangoeroe
+    print 'hallo ' dieren at random""")
+
+    result = hedy.transpile(code, self.level)
+
+    expected = textwrap.dedent("""\
+    dieren = ['Hond', 'Kat', 'Kangoeroe']
+    print(f'hallo {random.choice(dieren)}')""")
+
+    self.assertEqual(expected, result.code)
+    self.assertEqual(False, result.has_turtle)
+    self.assertIn(HedyTester.run_code(result), ['hallo Hond', 'hallo Kat', 'hallo Kangoeroe'])
+  def test_ask_print(self):
+
+    code = textwrap.dedent("""
     kleur is ask 'wat is je lievelingskleur?'
     print 'jouw lievelingskleur is dus' kleur '!'""")
 
@@ -119,274 +235,122 @@ class TestsLevel4(HedyTester):
 
     self.assertEqual(expected, result.code)
     self.assertEqual(False, result.has_turtle)
-  def test_print_if_else(self):
-    code = textwrap.dedent("""\
-    naam is Hedy
-    print 'ik heet' naam
-    if naam is Hedy print 'leuk' else print 'minder leuk'""")
+  def test_ask_assign(self):
+
+    code = textwrap.dedent("""
+    ding is kleur
+    kleur is ask 'Wat is je lievelings' ding
+    print 'Jouw favoriet is dus ' kleur""")
 
     result = hedy.transpile(code, self.level)
 
     expected = textwrap.dedent("""\
-    naam = 'Hedy'
-    print(f'ik heet{naam}')
-    if naam == 'Hedy':
-      print(f'leuk')
-    else:
-      print(f'minder leuk')""")
+    ding = 'kleur'
+    kleur = input('Wat is je lievelings'+ding)
+    print(f'Jouw favoriet is dus {kleur}')""")
 
     self.assertEqual(expected, result.code)
     self.assertEqual(False, result.has_turtle)
-  def test_print_if_else_ask(self):
 
+  def test_ask_assign_list(self):
     code = textwrap.dedent("""\
-    kleur is ask 'Wat is je lievelingskleur?'
-    if kleur is groen print 'mooi!' else print 'niet zo mooi'""")
-
-    result = hedy.transpile(code, self.level)
-
-    expected = textwrap.dedent("""\
-    kleur = input('Wat is je lievelingskleur?')
-    if kleur == 'groen':
-      print(f'mooi!')
-    else:
-      print(f'niet zo mooi')""")
-
-    self.assertEqual(expected, result.code)
-    self.assertEqual(False, result.has_turtle)
-  def test_print_if_else_with_line_break(self):
-    # line breaks should be allowed in if-elses until level 7 when we start with indentation
-    code = textwrap.dedent("""\
-    naam is Hedy
-    print 'ik heet' naam
-    if naam is Hedy print 'leuk'
-    else print 'minder leuk'""")
-
-    expected = textwrap.dedent("""\
-    naam = 'Hedy'
-    print(f'ik heet{naam}')
-    if naam == 'Hedy':
-      print(f'leuk')
-    else:
-      print(f'minder leuk')""")
+    color is gree, blue
+    choice is ask 'Is your favorite color one of: ' color""")
 
     self.multi_level_tester(
-      max_level=4,
       code=code,
-      expected=expected,
-      test_name=self.name(),
-      extra_check_function=self.is_not_turtle()
+      exception=hedy.exceptions.InvalidArgumentTypeException,
+      max_level=10,
+      test_name=self.name()
     )
-  def test_print_if_else_with_line_break_after_condition(self):
-    # line breaks after conditional should be allowed in if-elses until level 7 when we start with indentation
+
+  def test_forward_ask(self):
     code = textwrap.dedent("""\
-    naam is Hedy
-    print 'ik heet' naam
-    if naam is Hedy
-    print 'leuk'
-    else print 'minder leuk'""")
-
+    afstand is ask 'hoe ver dan?'
+    forward afstand""")
     expected = textwrap.dedent("""\
-    naam = 'Hedy'
-    print(f'ik heet{naam}')
-    if naam == 'Hedy':
-      print(f'leuk')
-    else:
-      print(f'minder leuk')""")
-
+    afstand = input('hoe ver dan?')
+    t.forward(afstand)
+    time.sleep(0.1)""")
     self.multi_level_tester(
-      max_level=4,
+      max_level=7,
       code=code,
       expected=expected,
-      test_name=self.name(),
-      extra_check_function=self.is_not_turtle()
+      extra_check_function=self.is_turtle(),
+      test_name=self.name()
     )
-  def test_if_else_newline_list_assigment_print(self):
-    # line breaks after conditional should be allowed in if-elses until level 7 when we start with indentation
-    code = textwrap.dedent("""\
-    people is mom, dad, Emma, Sophie
-    dishwasher is people at random
-    if dishwasher is Sophie
-    print 'too bad I have to do the dishes'
-    else
-    print 'luckily no dishes because' dishwasher 'is already washing up'""")
-
-    expected = textwrap.dedent("""\
-    people = ['mom', 'dad', 'Emma', 'Sophie']
-    dishwasher=random.choice(people)
-    if dishwasher == 'Sophie':
-      print(f'too bad I have to do the dishes')
-    else:
-      print(f'luckily no dishes because{dishwasher}is already washing up')""")
-
-    self.multi_level_tester(
-      max_level=4,
-      code=code,
-      expected=expected,
-      test_name=self.name(),
-      extra_check_function=self.is_not_turtle()
-    )
-
-
-  def test_print_if_else_line_break_and_space(self):
-    # line breaks should be allowed in if-elses until level 7 when we start with indentation
-
-    code = textwrap.dedent("""\
-    naam is Hedy
-    print 'ik heet' naam
-    if naam is Hedy print 'leuk'
-    else print 'minder leuk'""")
-
-    expected = textwrap.dedent("""\
-    naam = 'Hedy'
-    print(f'ik heet{naam}')
-    if naam == 'Hedy':
-      print(f'leuk')
-    else:
-      print(f'minder leuk')""")
-
-    self.multi_level_tester(
-      max_level=4,
-      code=code,
-      expected=expected,
-      test_name=self.name(),
-      extra_check_function=self.is_not_turtle()
-    )
-  def test_print_if_linebreak_statement(self):
-    # Breaking an if statement and its following statement should be
-    # permited until level 7 
-    
-    code = textwrap.dedent("""\
-    people is 1, 2, 3, 3
-    dishwasher is people at random
-    test is 1
-    if dishwasher is test
-    print 'too bad I have to do the dishes!'""")
-
-    expected = textwrap.dedent("""\
-    people = ['1', '2', '3', '3']
-    dishwasher=random.choice(people)
-    test = '1'
-    if dishwasher == test:
-      print(f'too bad I have to do the dishes!')""")
-
-    self.multi_level_tester(
-      max_level=4,
-      code=code,
-      expected=expected,
-      test_name=self.name(),
-      extra_check_function=self.is_not_turtle()
-    )
-  def test_print_if_else_and_var(self):
-
-    code = textwrap.dedent("""\
-    jouwkeuze is steen
-    computerkeuze is schaar
-    if computerkeuze is schaar and jouwkeuze is steen print 'jij wint'""")
-
-    result = hedy.transpile(code, self.level)
-
-    expected = textwrap.dedent("""\
-    jouwkeuze = 'steen'
-    computerkeuze = 'schaar'
-    if computerkeuze == 'schaar' and jouwkeuze == 'steen':
-      print(f'jij wint')""")
-
-    self.assertEqual(expected, result.code)
-    self.assertEqual(False, result.has_turtle)
-    self.assertEqual(HedyTester.run_code(result), 'jij wint')
-  def test_print_if_assign(self):
-    code = textwrap.dedent("""\
-    jouwkeuze is schaar
-    computerkeuze is schaar
-    if computerkeuze is jouwkeuze print 'gelijkspel!'""")
-
-    result = hedy.transpile(code, self.level)
-
-    expected = textwrap.dedent("""\
-    jouwkeuze = 'schaar'
-    computerkeuze = 'schaar'
-    if computerkeuze == jouwkeuze:
-      print(f'gelijkspel!')""")
-
-    self.assertEqual(expected, result.code)
-    self.assertEqual(False, result.has_turtle)
-    self.assertEqual(HedyTester.run_code(result), 'gelijkspel!')
-  def test_if_in_list(self):
-    code = textwrap.dedent("""\
-    items is red, green
-    selected is red
-    if selected in items print 'found!'""")
-
-    expected = textwrap.dedent("""\
-    items = ['red', 'green']
-    selected = 'red'
-    if selected in items:
-      print(f'found!')""")
-
-    result = hedy.transpile(code, self.level)
-
-    self.assertEqual(expected, result.code)
-    self.assertEqual(False, result.has_turtle)
-    self.assertEqual('found!', HedyTester.run_code(result))
-  # todo would be good to make combinations with if and turtle
-
 
 
   #negative tests
-  def test_indent_gives_parse_error(self):
+  def test_var_undefined_error_message(self):
     code = textwrap.dedent("""\
-    option is ask 'Rock Paper or Scissors?'
-    print 'Player 2 ' option
-    if option is Scissors
-        print 'Its a tie!'""")
+      naam is Hedy
+      print 'ik heet ' name""")
+
+    self.multi_level_tester(
+      code=code,
+      exception=hedy.exceptions.UndefinedVarException,
+      max_level=10,
+      test_name=self.name()
+    )
+
+    # deze extra check functie kan nu niet mee omdat die altijd op result werkt
+    # evt toch splitsen in 2 (pos en neg?)
+    # self.assertEqual('name', context.exception.arguments['name'])
+
+  def test_issue_375(self):
+    code = textwrap.dedent("""
+      is Foobar
+      print welcome""")
 
     with self.assertRaises(hedy.exceptions.ParseException) as context:
       result = hedy.transpile(code, self.level)
+
     self.assertEqual('Parse', context.exception.error_code)
-  def test_if_print_has_no_turtle(self):
+  def test_missing_opening_quote(self):
     code = textwrap.dedent("""\
-    jouwkeuze is schaar
-    computerkeuze is schaar
-    if computerkeuze is jouwkeuze print 'gelijkspel!'""")
-    result = hedy.transpile_inner(code, self.level)
-    self.assertEqual(False, result.has_turtle)
-  def test_no_space_after_keyword_gives_invalid(self):
-    code = textwrap.dedent("print'test'")
+      print hallo wereld'""")
 
-    self.multi_level_tester(
-      max_level=10,
-      code=code,
-      exception=hedy.exceptions.InvalidCommandException,
-      test_name=self.name()
-    )
-
-    #we don't have a function now for testing more exceptoion logic
-    # self.assertEqual('print', str(context.exception.arguments['guessed_command']))
-  def test_pront_should_suggest_print(self):
-    code = "pront 'Hedy is leuk!'"
-
-    with self.assertRaises(hedy.exceptions.InvalidCommandException) as context:
+    with self.assertRaises(hedy.exceptions.UnquotedTextException) as context:
       result = hedy.transpile(code, self.level)
-    self.assertEqual('Invalid', context.exception.error_code)
-    self.assertEqual('<span class="command-highlighted">print</span>', str(context.exception.arguments['guessed_command']))
-  def test_if_with_print_backtick(self):
+
+    self.assertEqual('Unquoted Text', context.exception.error_code)
+  def test_missing_all_quotes(self):
     code = textwrap.dedent("""\
-    name is ask 'ποιό είναι το όνομά σου;'
-    if name is Hedy print `ωραία` else print `μπου!`""")
+      print hallo wereld""")
 
     self.multi_level_tester(
-      max_level=4,
       code=code,
-      exception=hedy.exceptions.UnquotedTextException,
-      test_name=self.name()
+      max_level=4,
+      test_name=self.name(),
+      exception=hedy.exceptions.UndefinedVarException,
+    )
+  def test_print_Spanish(self):
+    code = textwrap.dedent("""\
+    print 'Cuál es tu color favorito?'""")
+    expected = textwrap.dedent("""\
+    print(f'Cuál es tu color favorito?')""")
+
+    self.multi_level_tester(
+      code=code,
+      max_level=11,
+      test_name=self.name(),
+      expected=expected,
+      extra_check_function=self.is_not_turtle()
     )
 
-  # def test_list_find_issue(self):
-  #   #'list' object has no attribute 'find'
-  #   # FH dd sept 2021 for later fixing!
-  #   code = textwrap.dedent("""\
-  #     নাম is ask আপনার নাম কি?
-  #     if নাম is হেডি print 'ভালো!' else print 'মন্দ'\"""")
-  #
-  #
+  #assorti
+  def test_detect_accented_chars(self):
+    self.assertEqual(True, hedy.hash_needed('éyyy'))
+    self.assertEqual(True, hedy.hash_needed('héyyy'))
+    self.assertEqual(False, hedy.hash_needed('heyyy'))
+
+
+
+
+
+
+
+
+
+
