+++ tests/test_level_11.py
@@ -1,4 +1,3 @@
-
 import hedy
 import textwrap
 from test_level_01 import HedyTester
@@ -6,312 +5,10 @@ from test_level_01 import HedyTester
 class TestsLevel11(HedyTester):
   level = 11
 
-  # ask tests
-  def test_ask_number_answer(self):
+  def test_if_with_indent(self):
     code = textwrap.dedent("""\
-            prijs is ask 'hoeveel?'
-            gespaard is 7
-            sparen is prijs - gespaard
-            print 'hallo' sparen""")
-    expected = textwrap.dedent("""\
-            prijs = input('hoeveel?')
-            try:
-              prijs = int(prijs)
-            except ValueError:
-              try:
-                prijs = float(prijs)
-              except ValueError:
-                pass
-            gespaard = 7
-            sparen = prijs - gespaard
-            print(f'hallo{sparen}')""")
-
-    self.multi_level_tester(
-      code=code,
-      expected=expected,
-      extra_check_function=self.is_not_turtle(),
-      test_name=self.name()
-    )
-
-  # new calculations
-  def test_int_addition_directly(self):
-    code = textwrap.dedent("""\
-            print 2 + 2""")
-    expected = textwrap.dedent("""\
-            print(f'{2 + 2}')""")
-
-    self.multi_level_tester(
-      code=code,
-      expected=expected,
-      extra_check_function=self.is_not_turtle(),
-      test_name=self.name()
-    )
-
-  def test_float_addition_directly(self):
-    code = textwrap.dedent("""\
-            print 2.5 + 2.5""")
-    expected = textwrap.dedent("""\
-            print(f'{2.5 + 2.5}')""")
-
-    self.multi_level_tester(
-      code=code,
-      expected=expected,
-      extra_check_function=self.is_not_turtle(),
-      test_name=self.name()
-    )
-
-  def test_float_addition_with_string(self):
-    code = textwrap.dedent("""\
-            print 'het antwoord is ' 2.5 + 2.5""")
-    expected = textwrap.dedent("""\
-            print(f'het antwoord is {2.5 + 2.5}')""")
-
-    self.multi_level_tester(
-      code=code,
-      expected=expected,
-      extra_check_function=self.is_not_turtle(),
-      test_name=self.name()
-    )
-
-  def test_float_addition_in_var(self):
-    code = textwrap.dedent("""\
-            antwoord is 2.5 + 2.5
-            print antwoord
-            """)
-    expected = textwrap.dedent("""\
-            antwoord = 2.5 + 2.5
-            print(f'{antwoord}')""")
-
-    self.multi_level_tester(
-      code=code,
-      expected=expected,
-      extra_check_function=self.is_not_turtle(),
-      test_name=self.name()
-    )
-
-  def test_float_addition_with_var(self):
-    code = textwrap.dedent("""\
-            getal1 is 5
-            getal2 is 4.3
-            print 'dat is dan: ' getal1 + getal2
-            """)
-    expected = textwrap.dedent("""\
-            getal1 = 5
-            getal2 = 4.3
-            print(f'dat is dan: {getal1 + getal2}')""")
-
-    self.multi_level_tester(
-      code=code,
-      expected=expected,
-      extra_check_function=self.is_not_turtle(),
-      test_name=self.name()
-    )
-
-  def test_store_calc_in_var(self):
-    code = textwrap.dedent("""\
-            a is 1
-            b is 2
-            c is a + b
-            print c ' is 3'""")
-    expected = textwrap.dedent("""\
-            a = 1
-            b = 2
-            c = a + b
-            print(f'{c} is 3')""")
-
-    self.multi_level_tester(
-      code=code,
-      expected=expected,
-      extra_check_function=self.is_not_turtle(),
-      test_name=self.name()
-    )
-
-  def test_add_string_vars(self):
-    code = textwrap.dedent("""\
-            getal1 is '5'
-            getal2 is '6'
-            print 'dat is dan: ' getal1 + getal2""")
-    expected = textwrap.dedent("""\
-            getal1 = '5'
-            getal2 = '6'
-            print(f'dat is dan: {getal1 + getal2}')""")
-
-    check_output = (lambda x: HedyTester.run_code(x) == 'dat is dan: 56')
-
-    self.multi_level_tester(
-      code=code,
-      expected=expected,
-      extra_check_function=check_output,
-      test_name=self.name()
-    )
-
-  def test_add_int_vars(self):
-    code = textwrap.dedent("""\
-            getal1 is 5
-            getal2 is 6
-            print 'dat is dan: ' getal1 + getal2""")
-    expected = textwrap.dedent("""\
-            getal1 = 5
-            getal2 = 6
-            print(f'dat is dan: {getal1 + getal2}')""")
-
-    check_output = (lambda x: HedyTester.run_code(x) == 'dat is dan: 11')
-
-    self.multi_level_tester(
-      code=code,
-      expected=expected,
-      extra_check_function=check_output,
-      test_name=self.name()
-    )
-
-  def test_assign_string_with_quotes(self):
-    code = textwrap.dedent("""\
-            name is 'felienne'
-            print name""")
-    expected = textwrap.dedent("""\
-            name = 'felienne'
-            print(f'{name}')""")
-    self.multi_level_tester(
-      code=code,
-      expected=expected,
-      extra_check_function=self.is_not_turtle(),
-      test_name=self.name()
-    )
-
-  def test_assign_string_with_quotes_and_string_value(self):
-    code = textwrap.dedent("""\
-            name is 'felienne'
-            print 'hallo ' name""")
-    expected = textwrap.dedent("""\
-            name = 'felienne'
-            print(f'hallo {name}')""")
-    self.multi_level_tester(
-      code=code,
-      expected=expected,
-      extra_check_function=self.is_not_turtle(),
-      test_name=self.name()
-    )
-
-  def test_print_multiple_calcs(self):
-    code = textwrap.dedent("""\
-            name is 1 + 2 + 3
-            print name""")
-
-    expected = textwrap.dedent("""\
-            name = 1 + 2 + 3
-            print(f'{name}')""")
-
-    self.multi_level_tester(
-      code=code,
-      expected=expected,
-      extra_check_function=self.is_not_turtle(),
-      test_name=self.name()
-    )
-
-  def test_calc_string_and_int(self):
-    code = textwrap.dedent("""\
-            x is 'test1'
-            y is x + 1""")
-
-    self.multi_level_tester(
-      code=code,
-      exception=hedy.exceptions.InvalidArgumentTypeException,
-      test_name=self.name()
-    )
-
-  def test_print_chained_assignments(self):
-    code = textwrap.dedent("""\
-            x is 1 + 2
-            y is x + 3
-            print y + 4""")
-
-    expected = textwrap.dedent("""\
-            x = 1 + 2
-            y = x + 3
-            print(f'{y + 4}')""")
-
-    self.multi_level_tester(
-      code=code,
-      expected=expected,
-      extra_check_function=self.is_not_turtle(),
-      test_name=self.name()
-    )
-
-  def test_assign_calc(self):
-    code = textwrap.dedent("""\
-            var is 5
-            print var + 5""")
-
-    expected = textwrap.dedent("""\
-            var = 5
-            print(f'{var + 5}')""")
-
-    self.multi_level_tester(
-      code=code,
-      expected=expected,
-      extra_check_function=self.is_not_turtle(),
-      test_name=self.name()
-    )
-
-  def test_access_variable_before_definition(self):
-    code = textwrap.dedent("""\
-            a is b
-            b is 3
-            print a""")
-
-    expected = textwrap.dedent("""\
-            a = b
-            b = 3
-            print(f'{a}')""")
-
-    self.multi_level_tester(
-      code=code,
-      expected=expected,
-      extra_check_function=self.is_not_turtle(),
-      test_name=self.name()
-    )
-
-  def test_list_creation_with_spaces(self):
-    code = textwrap.dedent("""\
-    actions is 'clap your hands', 'stomp your feet', 'shout Hurray'
-    for action in actions
-        for i in range 1 to 2
-            print 'if youre happy and you know it'
-            print action
-        print 'if youre happy and you know it and you really want to show it'
-        print 'if youre happy and you know it'
-        print action""")
-    expected = textwrap.dedent("""\
-    actions = ['clap your hands', 'stomp your feet', 'shout Hurray']
-    for action in actions:
-      step = 1 if int(1) < int(2) else -1
-      for i in range(int(1), int(2) + step, step):
-        print(f'if youre happy and you know it')
-        print(f'{action}')
-      print(f'if youre happy and you know it and you really want to show it')
-      print(f'if youre happy and you know it')
-      print(f'{action}')""")
-    result = hedy.transpile(code, self.level)
-
-    self.assertEqual(expected, result.code)
-    self.assertEqual(False, result.has_turtle)
-
-  def test_list_creation_with_numbers(self):
-    code = textwrap.dedent("""\
-    getallen is 1, 2, 3, 4, 5, 6, 7, 8, 9, 10
-    getal is getallen at random""")
-    expected = textwrap.dedent("""\
-    getallen = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
-    getal=random.choice(getallen)""")
-    result = hedy.transpile(code, self.level)
-
-    self.assertEqual(expected, result.code)
-    self.assertEqual(False, result.has_turtle)
-
-  def test_if_quotes_in_condition(self):
-    code = textwrap.dedent("""\
-    naam is 'Hedy'
-    if naam is 'Hedy'
+    naam is Hedy
+    if naam is Hedy
         print 'koekoek'""")
     expected = textwrap.dedent("""\
     naam = 'Hedy'
@@ -322,39 +19,7 @@ class TestsLevel11(HedyTester):
     self.assertEqual(expected, result.code)
     self.assertEqual(False, result.has_turtle)
 
-  def test_if_quotes_and_spaces_in_condition(self):
-    code = textwrap.dedent("""\
-    naam is 'Hedy is top'
-    if naam is 'Hedy is top'
-        print 'koekoek'""")
-    expected = textwrap.dedent("""\
-    naam = 'Hedy is top'
-    if str(naam) == str('Hedy is top'):
-      print(f'koekoek')""")
-    result = hedy.transpile(code, self.level)
-
-    self.assertEqual(expected, result.code)
-    self.assertEqual(False, result.has_turtle)
-
-  def test_if_else_quotes_in_condition(self):
-    code = textwrap.dedent("""\
-    naam is 'Hedy'
-    if naam is 'Hedy'
-        print 'koekoek'
-    else
-        print 'soepkip'""")
-    expected = textwrap.dedent("""\
-    naam = 'Hedy'
-    if str(naam) == str('Hedy'):
-      print(f'koekoek')
-    else:
-      print(f'soepkip')""")
-    result = hedy.transpile(code, self.level)
-
-    self.assertEqual(expected, result.code)
-    self.assertEqual(False, result.has_turtle)
-
-  def test_if_else_no_quotes_for_num_in_condition(self):
+  def test_if_else(self):
     code = textwrap.dedent("""\
     antwoord is ask 'Hoeveel is 10 plus 10?'
     if antwoord is 20
@@ -366,13 +31,6 @@ class TestsLevel11(HedyTester):
 
     expected = textwrap.dedent("""\
     antwoord = input('Hoeveel is 10 plus 10?')
-    try:
-      antwoord = int(antwoord)
-    except ValueError:
-      try:
-        antwoord = float(antwoord)
-      except ValueError:
-        pass
     if str(antwoord) == str('20'):
       print(f'Goedzo!')
       print(f'Het antwoord was inderdaad {antwoord}')
@@ -386,26 +44,174 @@ class TestsLevel11(HedyTester):
     self.assertEqual(False, result.has_turtle)
 
 
-  # negative tests
-  def test_assign_string_without_quotes(self):
-    code = textwrap.dedent("""\
-            name is felienne
-            print name""")
 
-    self.multi_level_tester(
-      code=code,
-      exception=hedy.exceptions.UnquotedAssignTextException,
-      test_name=self.name()
-    )
-  
-  def test_list_creation(self):
+  def test_for_loop(self):
     code = textwrap.dedent("""\
-    animals is 'duck', 'dog', 'penguin'""")
+    a is 2
+    a is 3
+    for a in range 2 to 4
+      a is a + 2
+      b is b + 2""")
     expected = textwrap.dedent("""\
-    animals = ['duck', 'dog', 'penguin']""")
+    a = '2'
+    a = '3'
+    step = 1 if int(2) < int(4) else -1
+    for a in range(int(2), int(4) + step, step):
+      a = int(a) + int(2)
+      b = int(b) + int(2)""")
+
     result = hedy.transpile(code, self.level)
 
     self.assertEqual(expected, result.code)
     self.assertEqual(False, result.has_turtle)
+
+  def test_if__else(self):
+    code = textwrap.dedent("""\
+    a is 5
+    if a is 1
+      x is 2
+    else
+      x is 222""")
+    expected = textwrap.dedent("""\
+    a = '5'
+    if str(a) == str('1'):
+      x = '2'
+    else:
+      x = '222'""")
+
+    result = hedy.transpile(code, self.level)
+
+    self.assertEqual(expected, result.code)
+    self.assertEqual(False, result.has_turtle)
+
+  def test_forloop(self):
+    code = textwrap.dedent("""\
+    for i in range 1 to 10
+      print i
+    print 'wie niet weg is is gezien'""")
+    expected = textwrap.dedent("""\
+    step = 1 if int(1) < int(10) else -1
+    for i in range(int(1), int(10) + step, step):
+      print(f'{i}')
+    print(f'wie niet weg is is gezien')""")
+
+    result = hedy.transpile(code, self.level)
+
+    self.assertEqual(expected, result.code)
+    self.assertEqual(False, result.has_turtle)
+
+
+  def test_reverse_range(self):
+    code = textwrap.dedent("""\
+    for i in range 10 to 1
+      print i
+    print 'wie niet weg is is gezien'""")
+    expected = textwrap.dedent("""\
+    step = 1 if int(10) < int(1) else -1
+    for i in range(int(10), int(1) + step, step):
+      print(f'{i}')
+    print(f'wie niet weg is is gezien')""")
+
+    result = hedy.transpile(code, self.level)
+
+    self.assertEqual(expected, result.code)
+    self.assertEqual(False, result.has_turtle)
+
+
+  def test_if_under_else_in_for(self):
+    code = textwrap.dedent("""\
+    for i in range 0 to 10
+      antwoord is ask 'Wat is 5*5'
+      if antwoord is 24
+        print 'Dat is fout!'
+      else
+        print 'Dat is goed!'
+      if antwoord is 25
+        i is 10""")
+
+    expected = textwrap.dedent("""\
+    step = 1 if int(0) < int(10) else -1
+    for i in range(int(0), int(10) + step, step):
+      antwoord = input('Wat is 5*5')
+      if str(antwoord) == str('24'):
+        print(f'Dat is fout!')
+      else:
+        print(f'Dat is goed!')
+      if str(antwoord) == str('25'):
+        i = '10'""")
+
+    result = hedy.transpile(code, self.level)
+
+    self.assertEqual(expected, result.code)
+    self.assertEqual(False, result.has_turtle)
+
+    #fails, issue 363
+
+  def test_for_ifbug(self):
+    code = textwrap.dedent("""\
+    for i in range 0 to 10
+      antwoord is ask 'Wat is 5*5'
+      if antwoord is 24
+        print 'fout'
+    print 'klaar met for loop'""")
+
+    expected = textwrap.dedent("""\
+      step = 1 if int(0) < int(10) else -1
+      for i in range(int(0), int(10) + step, step):
+        antwoord = input('Wat is 5*5')
+        if str(antwoord) == str('24'):
+          print(f'fout')
+      print(f'klaar met for loop')""")
+
+    result = hedy.transpile(code, self.level)
+
+    self.assertEqual(expected, result.code)
+    self.assertEqual(False, result.has_turtle)
+
+  def test_for_loopbug599(self):
+    code = textwrap.dedent("""\
+    for i in range 0 to 10
+      if i is 2
+        print '2'""")
+
+    expected = textwrap.dedent("""\
+      step = 1 if int(0) < int(10) else -1
+      for i in range(int(0), int(10) + step, step):
+        if str(i) == str('2'):
+          print(f'2')""")
+
+    result = hedy.transpile(code, self.level)
+
+    self.assertEqual(expected, result.code)
+    self.assertEqual(False, result.has_turtle)
+
+  def test_unindented_second_loop_1209(self):
+    code = textwrap.dedent("""\
+    for x in range 1 to 10
+     for y in range 1 to 10
+     print 'x*y'""")
+
+    with self.assertRaises(hedy.exceptions.NoIndentationException) as context:
+      result = hedy.transpile(code, self.level)
+
+  def test_dedented_second_loop_1209(self):
+    code = textwrap.dedent("""\
+    for x in range 1 to 10
+     for y in range 1 to 10
+    print 'x*y'""")
+
+    with self.assertRaises(hedy.exceptions.NoIndentationException) as context:
+      result = hedy.transpile(code, self.level)
+
+  def test_zigzag_indented_loop_1209(self):
+    code = textwrap.dedent("""\
+    for x in range 1 to 10
+      for y in range 1 to 10
+         print 'this number is'
+        print x*y""")
+
+    with self.assertRaises(hedy.exceptions.IndentationException) as context:
+      result = hedy.transpile(code, self.level)
+
